separate tests and move Auth::Page specific logic out of auth form integration test
