+++ ui/tests/helpers/general-selectors.ts
@@ -17,6 +17,7 @@ export const GENERAL = {
   secretTab: (name: string) => `[data-test-secret-list-tab="${name}"]`,
   flashMessage: '[data-test-flash-message]',
   latestFlashContent: '[data-test-flash-message]:last-of-type [data-test-flash-message-body]',
+  inlineAlert: '[data-test-inline-alert]',
 
   filter: (name: string) => `[data-test-filter="${name}"]`,
   filterInput: '[data-test-filter-input]',
